requirement.txt
PyPDF2==1.26.0
reportlab==3.5.42
python-docx==0.8.10
wikipedia==1.4.0
pillow==7.1.2
pyttsx3==2.87
PyAudio==0.2.11
SpeechRecognition==3.8.1